+++ android/src/main/java/ch/milosz/reactnative/RNZoomUsActivity.java
@@ -1,9 +1,9 @@
 package ch.milosz.reactnative;
 
 import us.zoom.sdk.ZoomSDK;
-import us.zoom.sdk.MeetingActivity;
+import us.zoom.sdk.NewMeetingActivity;
 
-public class RNZoomUsActivity extends MeetingActivity {
+public class RNZoomUsActivity extends NewMeetingActivity {
   @Override
   public void onBackPressed() {
     onClickLeave();